+++ DelvUI/Interface/PartyCooldowns/PartyCooldownsConfig.cs
@@ -170,7 +170,9 @@ namespace DelvUI.Interface.PartyCooldowns
                 Action? action = sheet?.GetRow(cooldown.ActionId);
                 if (action == null) { continue; }
 
-                if (action.Recast100ms > 0)
+                // get real cooldown from data
+                // keep hardcoded value for technical finish
+                if (action.Recast100ms > 0 && cooldown.ActionId != 16004)
                 {
                     cooldown.CooldownDuration = action.Recast100ms / 10;
                 }
@@ -363,7 +365,9 @@ namespace DelvUI.Interface.PartyCooldowns
             [118] = NewData(118, JobIDs.BRD, 50, 120, 15, 30, 3), // battle voice
             [7405] = NewData(7405, JobIDs.BRD, 88, 90, 15, 70, 2), // troubadour
             [7408] = NewData(7408, JobIDs.BRD, 66, 90, 15, 40, 2), // nature's minne
+            [25785] = NewData(25785, JobIDs.BRD, 90, 110, 15, 30, 3), // radiant finale
             [16012] = NewData(16012, JobIDs.DNC, 88, 90, 15, 70, 2), // shield samba
+            [16004] = NewData(16004, JobIDs.DNC, 70, 120, 20, 30, 3), // technical step / finish
             [16889] = NewData(16889, JobIDs.MCH, 56, 90, 15, 70, 2), // tactician
 
             // CASTER